Adding authorize in course and season methods in controller , adding gates in course and season policies

+++ modules/Course/Http/Controllers/SeasonController.php
@@ -6,20 +6,25 @@ use App\Http\Controllers\Controller;
 use Common\Responses\AjaxResponses;
 use Course\Http\Requests\SeasonRequest;
 use Course\Models\Season;
+use Course\Repositories\CourseRepo;
 use Course\Repositories\SeasonRepo;
 
 class SeasonController extends Controller
 {
     private $repository;
+    private $repository_course;
 
-    public function __construct(SeasonRepo $seasonRepo)
+    public function __construct(SeasonRepo $seasonRepo, CourseRepo $courseRepo)
     {
         $this->repository = $seasonRepo;
+        $this->repository_course = $seasonRepo;
     }
 
 
     public function store(SeasonRequest $request, $courseId)
     {
+        $course = $this->repository_course->findById($courseId);
+        $this->authorize('createSeason', $course);
         $this->repository->store($request, $courseId);
         newFeedback();
         return back();
@@ -28,11 +33,14 @@ class SeasonController extends Controller
     public function edit($seasonId)
     {
         $season = $this->repository->findById($seasonId);
+        $this->authorize('edit', $season);
         return view('Courses::seasons.edit', compact('season'));
     }
 
     public function update($seasonId, SeasonRequest $request)
     {
+        $season = $this->repository->findById($seasonId);
+        $this->authorize('edit', $season);
         $this->repository->update($seasonId, $request);
         // Find the course ID associated with the given season ID
         $courseId = $this->repository->findById($seasonId)->course_id;
@@ -43,13 +51,14 @@ class SeasonController extends Controller
     public function destroy($seasonId)
     {
         $season = $this->repository->findById($seasonId);
+        $this->authorize('delete', $season);
         $season->delete();
         return AjaxResponses::SuccessResponse();
     }
 
     public function accept($seasonId)
     {
-        //$this->authorize('change_confirmation_status', Season::class);
+        $this->authorize('change_confirmation_status', Season::class);
         if($this->repository->updateConfirmationStatus($seasonId, Season::CONFIRMATION_STATUS_ACCEPTED)){
             return AjaxResponses::SuccessResponse();
         }
@@ -58,7 +67,7 @@ class SeasonController extends Controller
 
     public function reject($seasonId)
     {
-        //$this->authorize('change_confirmation_status', Season::class);
+        $this->authorize('change_confirmation_status', Season::class);
         if($this->repository->updateConfirmationStatus($seasonId, Season::CONFIRMATION_STATUS_REJECTED)){
             return AjaxResponses::SuccessResponse();
         }
@@ -67,7 +76,7 @@ class SeasonController extends Controller
 
     public function lock($seasonId)
     {
-        //$this->authorize('change_confirmation_status', Season::class);
+        $this->authorize('change_confirmation_status', Season::class);
         if($this->repository->updateStatus($seasonId, Season::STATUS_LOCKED)){
             return AjaxResponses::SuccessResponse();
         }
@@ -76,7 +85,7 @@ class SeasonController extends Controller
 
     public function unlock($seasonId)
     {
-        //$this->authorize('change_confirmation_status', Season::class);
+        $this->authorize('change_confirmation_status', Season::class);
         if($this->repository->updateStatus($seasonId, Season::STATUS_OPENED)){
             return AjaxResponses::SuccessResponse();
         }
